requirements.txt
dateutils==0.6.6
netCDF4==1.2.4
numpy==1.11.1
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.0
six==1.10.0
